Test awnas pytorch models.

+++ fmodel.py
@@ -34,7 +34,7 @@ class EnsembleModel(DifferentiableModel):
         print("ensemble type : {};  number of model: {} ".format(self.ensemble_type, len(self.models)))
 
     def backward(self, gradient, image): # for foolbox 1.7.0 compatability
-        return np.sum(self.weights.reshape((len(self.models), 1, 1, 1)) * [model.backward(gradient, image) for model in self.models], axis=0)
+        return np.sum(self.weights.reshape((len(self.models), 1, 1, 1)) * [model.backward(gradient.astype(np.float32), image) for model in self.models], axis=0)
 
     def batch_predictions(self, images):
         preds = []
